+++ HHCServerManager/Models/ServerProfile.swift
@@ -223,6 +223,16 @@ struct NginxConfigSaveResult: Equatable, Hashable, Sendable {
     var capturedAt: Date
 }
 
+struct NginxConfigUpsertResult: Equatable, Hashable, Sendable {
+    var file: NginxConfigFile
+    var content: String
+    var backupPath: String?
+    var testResult: NginxTestResult
+    var createdNewFile: Bool
+    var rolledBack: Bool
+    var capturedAt: Date
+}
+
 enum FirewallBackend: String, CaseIterable, Identifiable, Sendable {
     case firewalld
     case ufw